requirements.txt
annotated-types==0.7.0
anyio==4.11.0
beautifulsoup4==4.13.5
Brotli==1.1.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
google-auth==2.40.3
google-genai==1.39.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
lxml==6.0.2
Markdown==3.9
mypy==1.18.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pluggy==1.6.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pytest==8.4.2
requests==2.32.5
rsa==4.9.1
ruff==0.13.2
sniffio==1.3.1
soupsieve==2.8
tenacity==9.1.2
types-beautifulsoup4==4.12.0.20250516
types-html5lib==1.1.11.20250917
types-Markdown==3.9.0.20250906
types-requests==2.32.4.20250913
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
ruff==0.6.2
mypy==1.11.1
